Moved Model interface to separate namespace
