+++ DocxTemplater.Test/ModelPropertyAttributeTest.cs
@@ -2,6 +2,7 @@
 using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Wordprocessing;
+using DocxTemplater.Model;
 
 namespace DocxTemplater.Test
 {
