+++ DocxTemplater/Formatter/VariableReplacer.cs
@@ -1,5 +1,6 @@
 ﻿using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Wordprocessing;
+using DocxTemplater.Model;
 using System;
 using System.Collections.Generic;
 using System.Linq;
