+++ DocxTemplater/ModelLookup.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using DocxTemplater.Model;
+using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
